Test/requirements.txt
pytest==8.3.2
requests==2.32.3
gitpython==3.1.43
jsonschema==4.23.0
tenacity==9.0.0(venv)
